fix: 우리동네 관련 변경된 API 수정 및 UI 로직 수정

+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/TogetherViewModel.kt
@@ -4,11 +4,11 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 
 interface TogetherViewModel {
-    val filterRequiredAmount: StateFlow<Int>
-    val filterMaxMember: StateFlow<Int>
+    val filterRequiredAmount: StateFlow<Long>
+    val filterMaxMember: StateFlow<Long>
     val showDialog: SharedFlow<Boolean>
 
     fun onFilterClick()
 
-    fun getFilters(requiredAmount: Int, maxMember: Int)
+    fun getFilters(requiredAmount: Long, maxMember: Long)
 }
